fix: fix up project folders and linter
