--- packages/website/server/@qwik-city-static-paths.js
@@ -1,27 +1,0 @@
-const staticPaths = new Set([]);
-function isStaticPath(method, url) {
-  if (method.toUpperCase() !== 'GET') {
-    return false;
-  }
-  const p = url.pathname;
-  if (p.startsWith('/build/')) {
-    return true;
-  }
-  if (p.startsWith('/assets/')) {
-    return true;
-  }
-  if (staticPaths.has(p)) {
-    return true;
-  }
-  if (p.endsWith('/q-data.json')) {
-    const pWithoutQdata = p.replace(/\/q-data.json$/, '');
-    if (staticPaths.has(pWithoutQdata + '/')) {
-      return true;
-    }
-    if (staticPaths.has(pWithoutQdata)) {
-      return true;
-    }
-  }
-  return false;
-}
-export { isStaticPath };
